WIP: UTIL_EST: use estimated utilization on load balancing paths
Load balancing paths are triggered when we are outside of the energy-aware
more or by the periodic balancer. In these cases, the usage of an
estimated utilization for the available capacity on candidate CPUs,
can provide a more stable information to refine the scheduling decisions.

This patch updates all the slow paths to use the estimated utilization
whenever possible, i.e. when the UTIL_EST feature is enabled.
Specifically we use the estimated CPU utilization to support:
a) identification of the idlest group
   via find_idlest_group=>task_fits_spare
b) identification of the idlest CPU
   via find_idlest_cpu=>task_fits_spare

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: infinity-plus <navdhareharsh2001@gmail.com>